Fix input being broken when a search is running
Due to how Preact handles onChange the state would not be updated until
the input was blurred, this made it impossible to enter a new search
while polling was in progress because the poller would continuously
trigger a new render. Fixed by using onInput instead. It is unclear why
preact and react behave differently here.