Use pointers in JobRepository maps
The primary use of this is updating NumMatchedEvents in stats without
needing to write to the map. I removed the write to the stats map in a
recent commit but forgot to change the map to use pointers. Without this
change, the server can panic due to concurrent reading and writing of
the stats map.

There are probably still a few other cases where the server will panic
due to concurrent map read/writes.