Use FTS "MATCH" keyword to filter at SQLite level
Combined with using FTS4 and setting order=DESC, this results in a
massive speedup. Setting order=DESC alone gave a 9x speedup on my
current testing file (70MB from logdunk).

Still unsure if this is fast enough. The query planner seems to search
the entire raws table before filtering by source and timestamp. But
maybe it's not possible to get it to do it the other way around?