Prevent duplicate events in SQLite mode
Before this change, you would have duplicate events stored in the
database every time you restarted the program.

The SQLite EventRepository now has a UNIQUE constraint on the
combination of source, timestamp and offset in the file.

In theory this could end up dropping legitimate events in a
"rolling file" scenario if there was an event at the exact offset in the
new file, but since the timestamp should also be extracted from the
event and is also compared against this should be a non-issue.

There may be a configuration flag in the future that will allow you to
skip this check.