Move job logic from web, fix state bug
Because of an oopsie in the order of arguments when updating the state
of a job in SQLite jobs would get stuck in a running state.

The job logic being moved from web is a hygiene thing that will
hopefully make it easier to maintain and add features in the long run.