Try to fix working dir problem in CI
When running locally there is no problem with doing "go generate ./cmd/logsuck/main.go" which refers to "../../web/static/dist" (which is ./web/static/dist from the project root which definitely exists. But when running in GH Actions this does not work for some reason, I suspect the working directory is not being set to ./cmd/logsuck when running packager.go.