#13 Refactor how search is parsed
This commit lays the groundwork for adding more complex search commands,
such as rex, where, stats, etc.

This introduces the concepts of pipelines. A pipeline is a series of
steps, where a step takes a slice of events as input and generates a new
slice of events as output. The only step currently is search, which
ignores its input and retrieves its output from the event repository.

Each pipeline step runs in its own goroutine and receives an input
channel and an output channel. The step is expected to close the output
channel when it is done or when the context is cancelled.

This commit also fixes #15.