Replace net/http with Gin gonic
The short term benefit of this change is to reduce the size of the
web backend code. ~80 LOC out of 300 were removed thanks to
serialization / deserialization being easier.

Long term I think this will pay dividends as the web parts become more
complex with authentication and authorization etc.

Binary size increased by about 50% though (based on very un-scientific
testing) which is a lot. For now it's worth the tradeoff but it's
something to keep an eye on.