Add "true batch" mode
This increases throughput of adding new events by 2-3x.

Previously the repository would add events one by one, making two
queries per event. It was done this way to handle duplicates and to
ensure the EventRaws row had the same ID as the corresponding Events row

The new implementation instead does three queries for the entire batch.
The way this works is using AUTOINCREMENT + ON CONFLICT IGNORE. When
there is a duplicate in the batch, it will be ignored by ON CONFLICT
IGNORE, but because of AUTOINCREMENT the ID will still be "used up",
leaving a hole in the sequence of IDs. After all the non-duplicate
events have been inserted into Events, ALL events are inserted into
EventRaws. Since the IDs in the Events table are used up even on
duplicate events, the ID sequence of EventRaws should still line up with
the Events sequence. After both inserts are performed, the "garbage"
events in EventRaws are removed by deleting all EventRaws rows which do
not have a corresponding Events row.

This is way faster, but I am not sure it is robust in all situations and
the AUTOINCREMENT + ON CONFLICT IGNORE behavior seems like something
that might change in a future version of SQLite? The other drawback is
that handling duplicates is now way slower, since the EventRaws need to
be added and then removed now whereas before they would just try to add
to the Events table and then skip the raw after getting a conflict.
My thinking is that trying to add duplicates should be the exception, so
speeding up the common case is worth the tradeoff.

True batch can be disabled by setting "sqlite.trueBatch" to false in the
configuration file.