Add SourceId to events
A SourceId represents a specific "generation" of a file. If a file is
rolled, a new SourceId should be generated. This makes it easier to sort
and filter events that have the same source but are actually from
different generations.

A problem with implementing #24 has been that simply finding nearby
events using source + offset + timestamp could result in events from a
different generation interfering with the results, ,which this should
solve.

This is a breaking change, so logsuck.db needs to be deleted before
using a version of Logsuck with this commit included.