Use sass and CSS modules
Sass because I'm used to it. CSS modules because I wanted the CSS to be
scoped to each component.

Unfortunately there is no autocompletion when importing classes from
sass into the TypeScript files. I found a plugin for it but it
increased the npm dependency count of this project by about 50% which
is not acceptable.

There are still some global styles in style.css. These are mostly
utility classes. The styling for the navbar is still there as well
since the navbar may be shown even when the JS bundle does not load.