Preserve column order when using | table
Previously the columns would be in an arbitrary order based on iterating
over the map of values on the first row.

Now the column order is saved to the database when creating the job and
used to show the columns in the order they were given in the pipeline.

This adds a new column `column_order_json` to the Jobs table, so it
requires a migration or recreating the database.