WIP Plugin support
The idea:
* Move public interfaces to pkg/logsuck. There shouldn't be much code
here, just structs and interfaces. Packages in this folder must not
use internal packages. Packages in this folder are used by plugins.

* Put plugins in the plugins directory. Plugins should not use any
internal packages. Plugins implement interfaces and provide them into
the dependency injection context on startup.

* Internal packages can reference pkg/logsuck freely. Internal packages
should not reference plugins. The only exceptions are the
DependencyInjection.go file and tests which may use plugins for
convenience.

Plugins are currently not loadable dynamically. Instead, if you want to
use a plugin you must update the DependencyInjection.go file and build.
This seems like an acceptable tradeoff for now since there are some
caveats with using the Go plugin system.