Generate config schema dynamically
This allows plugins and tasks to provide schemas for their configuration
, making it much easier to configure them through the GUI.

A downside of this is that SQLite configuration is moved to the
"plugins" key in the config. Since SQLite is supposed to be a part of
the base Logsuck build this is a bit ugly. Maybe SQLite should be
special cased to also be included in the base configuration.