Add the concept of "core plugins"
Core plugins are treated specially when it comes to the config schema.
Instead of being stored under the plugins key in the schema, core
plugins can put their configuration in the root of the config object.
This keeps compatibility with Logsuck configs created before plugins
were added and looks much cleaner in the schema.